v1.0.0

+++ mall-business/mall-business-system/src/main/java/com/emond/mall/business/system/service/impl/UserServiceImpl.java
@@ -3,18 +3,15 @@ package com.emond.mall.business.system.service.impl;
 
 import com.emond.mall.business.system.domain.Role;
 import com.emond.mall.business.system.domain.User;
-import com.emond.mall.business.system.domain.query.UserQueryCriteria;
 import com.emond.mall.business.system.repository.UserRepository;
 import com.emond.mall.business.system.service.RoleService;
 import com.emond.mall.business.system.service.UserService;
 import com.emond.mall.common.exception.BadRequestException;
 import com.emond.mall.common.exception.EntityExistException;
-import com.emond.mall.common.exception.ResourceNotFoundException;
+import com.emond.mall.common.service.impl.BaseServiceImpl;
 import com.emond.mall.common.utils.OAuth2Utils;
-import com.emond.mall.common.utils.QueryHelp;
+import com.emond.mall.provider.system.dto.UserPassDTO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -27,17 +24,27 @@ import java.util.stream.Collectors;
 
 @Service
 @Transactional(propagation = Propagation.SUPPORTS, readOnly = true, rollbackFor = Exception.class)
-public class UserServiceImpl implements UserService {
+public class UserServiceImpl extends BaseServiceImpl<User> implements UserService {
+
+    private final static String NAME = "角色";
+    private final UserRepository userRepository;
+
     @Autowired
-    private UserRepository userRepository;
+    public UserServiceImpl(UserRepository userRepository) {
+        super(userRepository, NAME);
+        this.userRepository = userRepository;
+    }
+
     @Autowired
     private PasswordEncoder passwordEncoder;
+
     @Autowired
     private RoleService roleService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public User create(User resources) {
+        checkLevel(resources);
         if (userRepository.existsByUsername(resources.getUsername())) {
             throw new EntityExistException("账号", resources.getUsername());
         }
@@ -50,14 +57,18 @@ public class UserServiceImpl implements UserService {
         }
         resources.setPassword(passwordEncoder.encode("123456"));
 
-        return userRepository.save(resources);
+        return super.create(resources);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(Set<Long> ids) {
         for (Long id : ids) {
-            checkLevel(id);
+            Integer currentLevel = Collections.min(roleService.findByUsersId(OAuth2Utils.getCurrentUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
+            Integer optLevel = Collections.min(roleService.findByUsersId(id).stream().map(Role::getLevel).collect(Collectors.toList()));
+            if (currentLevel > optLevel) {
+                throw new BadRequestException("角色权限不足，不能操作");
+            }
             userRepository.deleteById(id);
         }
     }
@@ -65,29 +76,22 @@ public class UserServiceImpl implements UserService {
     /**
      * 如果当前用户的角色级别低于创建用户的角色级别，则抛出权限不足的错误
      *
-     * @param userId 被操作用户ID
+     * @param resource 被操作用户
      */
-    private void checkLevel(Long userId) {
+    private void checkLevel(User resource) {
         Integer currentLevel = Collections.min(roleService.findByUsersId(OAuth2Utils.getCurrentUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
-        Integer optLevel = Collections.min(roleService.findByUsersId(userId).stream().map(Role::getLevel).collect(Collectors.toList()));
+        Integer optLevel = Collections.min(resource.getRoles().stream().map(Role::getLevel).collect(Collectors.toList()));
         if (currentLevel > optLevel) {
             throw new BadRequestException("角色权限不足，不能操作");
         }
     }
 
-    @Override
-    public User findById(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("用户", "ID", id));
-        return user;
-    }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public void update(User resources) {
+    public User update(User resources) {
         Long userId = resources.getId();
-        this.findById(userId);
-        checkLevel(userId);
+        checkLevel(resources);
         if (userRepository.existsByUsernameAndIdNot(resources.getUsername(), userId)) {
             throw new EntityExistException("账号", resources.getUsername());
         }
@@ -98,15 +102,26 @@ public class UserServiceImpl implements UserService {
         if (userRepository.existsByPhoneAndIdNot(resources.getPhone(), userId)) {
             throw new EntityExistException("电话", resources.getPhone());
         }
-        userRepository.save(resources);
+        return super.update(resources);
     }
 
     @Override
-    public Page<User> getUserPage(UserQueryCriteria criteria, Pageable pageable) {
+    public void updatePass(UserPassDTO userPassDTO) {
+        User currentUser = this.findById(OAuth2Utils.getCurrentUserId());
+        if(!passwordEncoder.matches(userPassDTO.getOldPass(), currentUser.getPassword())){
+            throw new BadRequestException("修改失败，旧密码错误");
+        }
+        if(passwordEncoder.matches(userPassDTO.getNewPass(), currentUser.getPassword())){
+            throw new BadRequestException("新密码不能与旧密码相同");
+        }
 
-        return userRepository.findAll((root, criteriaQuery, criteriaBuilder)
-                -> QueryHelp.getPredicate(root,criteria,criteriaBuilder),pageable);
+        currentUser.setPassword(passwordEncoder.encode(userPassDTO.getNewPass()));
+        super.update(currentUser);
     }
 
 
+    @Override
+    public void updateProfile(User resources) {
+        super.update(resources);
+    }
 }
